mosh: sms: Increase shell buffer for long SMS.
Shell command buffer is limiting the length of the
SMS that can be sent. CONFIG_SHELL_CMD_BUFF_SIZE
is now increased from default 256 to 320 to allow
SMS messages at least up to 270 characters.
The actual length depends on how sms send command
is used with short vs. long options and spaces.
With a number of 12 characters, maximum message
length after the following command is 293:
    sms send -n123456789012 -m

Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>